+++ aws_ecs_service/package/cfnresponse.py
@@ -1,26 +1,22 @@
-#  Copyright 2016 Amazon Web Services, Inc. or its affiliates. All Rights Reserved.
-#  This file is licensed to you under the AWS Customer Agreement (the "License").
-#  You may not use this file except in compliance with the License.
-#  A copy of the License is located at http://aws.amazon.com/agreement/ .
-#  This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, express or implied.
-#  See the License for the specific language governing permissions and limitations under the License.
-
+import logging
 import json
 
 from botocore.vendored import requests
 
 SUCCESS = "SUCCESS"
 FAILED = "FAILED"
+MAXIMUM_RESPONSE_SIZE = 1024 * 4
 
 
-def send(event, context, responseStatus, responseData, physicalResourceId=None, noEcho=False):
+def send(event, context, responseStatus, responseData, physicalResourceId=None, noEcho=False, reason=None):
     responseUrl = event['ResponseURL']
 
-    print(responseUrl)
+    logging.info(f'Sending response back to CloudFormation: {responseUrl}...')
 
     responseBody = {}
+
     responseBody['Status'] = responseStatus
-    responseBody['Reason'] = 'See the details in CloudWatch Log Stream: ' + context.log_stream_name
+    responseBody['Reason'] = reason or ('See the details in CloudWatch Log Stream: ' + context.log_stream_name)
     responseBody['PhysicalResourceId'] = physicalResourceId or context.log_stream_name
     responseBody['StackId'] = event['StackId']
     responseBody['RequestId'] = event['RequestId']
@@ -28,20 +24,25 @@ def send(event, context, responseStatus, responseData, physicalResourceId=None, 
     responseBody['NoEcho'] = noEcho
     responseBody['Data'] = responseData
 
-    json_responseBody = json.dumps(responseBody, default=lambda o: '<not serializable>')
+    response_json = json.dumps(responseBody, default=lambda o: '<not serializable>')
+    is_too_big = len(response_json.encode('utf-8')) >= MAXIMUM_RESPONSE_SIZE
 
-
-    print("Response body:\n" + json_responseBody)
+    if is_too_big:
+        responseBody['Data'] = None
+        response_json = json.dumps(responseBody, default=lambda o: '<not serializable>')
 
     headers = {
         'content-type': '',
-        'content-length': str(len(json_responseBody))
+        'content-length': str(len(response_json))
     }
 
     try:
-        response = requests.put(responseUrl,
-                                data=json_responseBody,
-                                headers=headers)
-        print("Status code: " + response.reason)
+        response = requests.put(
+            responseUrl,
+            data=response_json,
+            headers=headers
+        )
+
+        logging.info("Status code: " + response.reason)
     except Exception as e:
-        print("send(..) failed executing requests.put(..): " + str(e))
+        logging.info("send(..) failed executing requests.put(..): " + str(e))