+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Release history
 
+#### 1.0.6
+Various bugs fixed related to custom resource. Added more logging
+for better debugging. Improved kwargs fixer. Improved cfn 
+response sender.
+
 #### 1.0.5
 Implement a fail-safe for json dumps.